pos_backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.1
mysqlclient>=2.2
Pillow>=10.0
qrcode>=7.4
reportlab>=4.0
openpyxl>=3.1
django-filter>=23.0
